+++ src/photos/components/Lightbox/BottomNavigation.tsx
