+++ src/photos/components/Lightbox/Buttons.tsx
