+++ src/photos/components/Lightbox/MainImage.tsx
