+++ src/photos/components/Lightbox/index.tsx
@@ -1,6 +1,6 @@
-import BottomNavigation from '@src/photos/components/single/Lightbox/BottomNavigation'
-import Buttons from '@src/photos/components/single/Lightbox/Buttons'
-import MainImage from '@src/photos/components/single/Lightbox/MainImage'
+import BottomNavigation from '@src/photos/components/Lightbox/BottomNavigation'
+import Buttons from '@src/photos/components/Lightbox/Buttons'
+import MainImage from '@src/photos/components/Lightbox/MainImage'
 import { range } from '@src/photos/utils/range'
 import type { ImageProps, SharedModalProps } from '@src/photos/utils/types'
 import { MotionConfig } from 'framer-motion'
